backend/requirements_minimal.txt
Flask
Flask-CORS
python-dotenv
requests
opencv-python
Pillow
numpy